requirements.txt
# Core scientific stack
xarray
numpy
pandas
netCDF4

# xclim for computing climate indicators
xclim

# LangChain and AI agent framework
langchain
langchain-openai
langchain-chroma
langchain-community
langgraph
langsmith

# Embeddings & OpenAI
openai
python-dotenv  # to load API keys from a .env file

# Pydantic for dynamic validation
pydantic>=2.0

# YAML utilities
pyyaml

# Graphs and visualization
matplotlib
seaborn

bs4
openmeteo_requests
pymannkendall